Fixed pie

+++ math/include/Math/Constants.h
@@ -7,7 +7,7 @@
 #define MATH_CONSTANTS_H_INCLUDED
 namespace Math
 {
-    constexpr unsigned pie = 3.14159265358979323846;
+    constexpr double pie = 3.14159265358979323846;
     constexpr unsigned tau = pie * 2;
 }
 #endif // CONSTANTS_H_INCLUDED
